Resolve conflict
