+++ torchaudio/csrc/sox/effects.cpp
@@ -1,7 +1,7 @@
 #include <sox.h>
-#include <torchaudio/csrc/sox/sox_effects.h>
-#include <torchaudio/csrc/sox/sox_effects_chain.h>
-#include <torchaudio/csrc/sox/sox_utils.h>
+#include <torchaudio/csrc/sox/effects.h>
+#include <torchaudio/csrc/sox/effects_chain.h>
+#include <torchaudio/csrc/sox/utils.h>
 
 using namespace torchaudio::sox_utils;
